Struct abort_on_panic::PanicGuard

pub struct PanicGuard {
    // some fields omitted
}

Once this object is created, it can only be destroyed in an orderly fashion. Attempting to clean it up from a panic handler will abort the process.

Methods

impl PanicGuard

fn new() -> PanicGuard

Create a panic guard with a generic message.

fn with_message(message: &'static str) -> PanicGuard

Create a panic guard with a custom message.